Converted authjs to typescript
